updated classification algorithm
taxonomic data is now stored in the tree nodes
this allows for classifications directly
from the nearestNeighbor data objects in VPTree

The classification now weights each class
with 1/distance from that class.

One can select if only the highest weight class is to be chosen
or if the whole weight matrix should be returned.

added a file createVPTree that creates a vptree and saves it to file
and subsetmetadata as help functions/classes to classify.py
but they can also be used in other places